+++ internal/dir.h
@@ -1,0 +1,16 @@
+#ifndef INTERNAL_DIR_H /* -*- C -*- */
+#define INTERNAL_DIR_H
+/**
+ * @file
+ * @brief      Internal header for Dir.
+ * @author     \@shyouhei
+ * @copyright  This  file  is   a  part  of  the   programming  language  Ruby.
+ *             Permission  is hereby  granted,  to  either redistribute  and/or
+ *             modify this file, provided that  the conditions mentioned in the
+ *             file COPYING are met.  Consult the file for details.
+ */
+
+/* dir.c */
+VALUE rb_dir_getwd_ospath(void);
+
+#endif /* INTERNAL_DIR_H */
