+++ internal/loadpath.h
@@ -1,0 +1,17 @@
+#ifndef INTERNAL_LOADPATH_H /* -*- C -*- */
+#define INTERNAL_LOADPATH_H
+/**
+ * @file
+ * @brief      Internal header for $LOAD_PATH.
+ * @author     \@shyouhei
+ * @copyright  This  file  is   a  part  of  the   programming  language  Ruby.
+ *             Permission  is hereby  granted,  to  either redistribute  and/or
+ *             modify this file, provided that  the conditions mentioned in the
+ *             file COPYING are met.  Consult the file for details.
+ */
+
+/* loadpath.c */
+extern const char ruby_exec_prefix[];
+extern const char ruby_initial_load_paths[];
+
+#endif /* INTERNAL_LOADPATH_H */
