+++ internal/range.h
@@ -1,0 +1,18 @@
+#ifndef INTERNAL_RANGE_H /* -*- C -*- */
+#define INTERNAL_RANGE_H
+/**
+ * @file
+ * @brief      Internal header for Range.
+ * @author     \@shyouhei
+ * @copyright  This  file  is   a  part  of  the   programming  language  Ruby.
+ *             Permission  is hereby  granted,  to  either redistribute  and/or
+ *             modify this file, provided that  the conditions mentioned in the
+ *             file COPYING are met.  Consult the file for details.
+ */
+
+/* range.c */
+#define RANGE_BEG(r) (RSTRUCT(r)->as.ary[0])
+#define RANGE_END(r) (RSTRUCT(r)->as.ary[1])
+#define RANGE_EXCL(r) (RSTRUCT(r)->as.ary[2])
+
+#endif /* INTERNAL_RANGE_H */
